Fix history lookup from insert mode
Previously, the only way to get history lookup (with fzf) was `<esc>/`
which would go to normal mode with `<esc>`, then do `/`, which is bound
to history lookup.

History lookup with a prefix used LBUFFER to get the command you started
to type, and pre-fill the command lookup search with it.

Going into normal mode moved the cursor 1 char to the left, and the
prefix was 1 char shorter.
So when I write 1 char, then do an history lookup, I had no prefix.

By using an explicit keybind in insert mode to trigger an history
lookup, the LBUFFER is preserved, and 1 char-prefix works OK.