nvim: Fix 'epic' slowness in (big) vim files
This was because the Colorizer plugin (a vimscript one) was triggering
itself on sooo many autocommands! like CursorMoved in normal mode
(but whyyy??? no idea..), and in TextChangedI, CursorMovedI.

Now editting a vim file is fast again, or at least not slow as hell like
it was before this change! I still have a bit of lag, but it's def' usable!