wezterm: !FIX! Ctrl-click for links, mouse-aware term apps are unfucked!
(from sg config)

The issue was that Only Ctrl-click >>Up<< event was mapped by wezterm, so
the 'Down' event was sent to the running program, but the Up event
was NEVER sent!
And so I always ended up with weird behavior in tmux, in neovim, ...

The solution is to simply disable Ctrl-click >>Down<< event, so that it
is handled by wezterm to 'do nothing', and the Up event will be properly
used to open links \o/

I am happy!