zsh: fzf: !!! Fix prompt rendering and off-by-one redraw issue <3
The prompt rendering issue was because I had a _local_ variable named
`prompt` which was actually overwriting the builtin one in zsh.
Hitting enter would re-show it because I have different prompts for the
current & past commands, and the `prompt` var is overwriten at this
moment.
=> So it was not that the reset-prompt didn't work, just that I didn't
   have a prompt until it was reset when ending/starting a new cmdline.

The off-by-one redraw issue, making the prompt move by 1 line each time
the fzf zwidget was shown, was because I had a 'debug' using `zle -M`
showing the returned results... BEFORE the `zle reset-prompt`, making it
start on the _wrong_ line.

Now I feel stupid.. the off-by-one error was due to a debug..
=> I think this is an acheivement of some kind ^^