ahk: Add Win+² to activate last flashing window (needs system plumbing!)
Now I don't have to search for that flashing window anymore / nor have
to use the mouse! (or count the index of the taskbar icon)

When I get a new message... just press Win+² and the window is
activated/raised/focused! And if my other keys I can easily move it to
the position I want, like the other monitor with Win+Shift+s, or center
it on monitor with mouse with Win+Ctrl+Space \o/

---
This is one of the few AHK code to start doing 'low level' stuff with
the system, here I'm registering an AHK function to all Shell events
(the desktop environment) to monitor global messages (nice!).