zsh: Fix lost-on-reboot gitstatusd binary & selected syntax theme
gitstatus' bootstrap was auto-downloading its pre-compiled binary to
a directory in ~/.cache, which is a tmpfs => empty on boot.

Similarly fast-syntax-highlighting was saving which theme is selected in
a directory in ~/.cache => empty on boot, so it was falling back on
default theme.
(also update this plugin, and point it to an updated fork)