nvim-next: Fix statusline: `~/foo` & `./foo` when file in CWD
`~/foo` was borked, it would show the tilde for every file, since all
basically started with /home/myuser..
Now only shows the tilde when the parent_path is exactly /home/myuser.

A file in CWD would always show the parent dir name, instead of `.` when
file in CWD, now fixed!