Nix,zsh: Initial nixification of zsh config! \o/
Adds many configuration points, that are (currently) 'harshly'
overwritten from Nix when packaging the config.

Config is NOT editable. BUT if _should_ still work if linked as before
on a non-NixOS system.

Now plugins are declared and locked in Nix (not yet in a flake), and
binary dependencies that made sense have been identified and provided :)

(mpv & ffmpeg binary deps don't make sense to have for a shell definition)
(helper functions and aliases are NOT pure, they reference potentially
unavailable software. This will probabaly be reworked / made more
flexible / more configurable with time)