nvim-next: Add system to change what is returned by `vim.fn.stdpath`
This is necessary to ensure the nvim configuration is self-contained.
It allows my configs and other plugins that want to refer to the config
dir (for example) refers to my custom config dir, not the default one in
~/.config/nvim.

It's not perfect though, as vimscript's `stdpath` is NOT overriden, only
the Lua one (`vim.fn.stdpath`).