nvim-next: LUA REWRITE: init, plugins, add declarative plugins system
Config can load!
Although not fully bootstrapped as lazy.nvim needs to be available in a
pack/ in packpath

ADD a declarative plugin framework, plugins can refer to other plugin
via a unique ID, plugins can be tagged.

One of the plugin is `lazy.nvim`, with a `on_boot` field that describes
how to boot my plugins with `lazy.nvim` ❤️ ❤️ ❤️.
=> The boot system (triggered in `init.lua`) is not even aware of which
   package manager is used!

I want to extend this further to use tags system to change how the
plugin is loaded / when to acheive fast minimal initial boot, with
plugins activating as needed in an abstract way. #yolo