zsh: prompt: Better match git colors and make segment more readable
In git status: staged is green and modified is red

In zsh git segment:
Before: staged was red (`+`) and modified was yellow (`!`)
After:  staged is green (`+`) and modified is red (`~`)

The change of symbol for modified feels simpler, more logical.
In my nvim config I already have `~` for modified lines, and `!` really
felt like something is wrong (and even more so when modified is red not
yellow), which isn't right.
Now red `~` feels right, there are changes, they're not committed yet.

And the branch name is now the only thing that can be yellow, showing
that there are modified/staged files in the worktree.