nvim-next: Always show the byte-column by default
The virtual column was weird when cursor is on a wrapped line and
'linebreak' is set, as the virtual-column is actually more the
screen-column, where a wrapped line with showbreak or linebreak set will
add useless columns to the counter.

The virtual-column is still useful when 'virtualedit' is set, so it is
enabled (as additional info) in this case.