nvim-next: Add direct window focus by winnr
All non-current windows that are known to be 'replaceable'
(neotree window tries hard not to be replaced for example) now have
their winnr in the statusline.
This winnr is replacing the Mode in the statusline which was especially
not useful for non-current window.

Also adds <C-w>N where (1 <= N <= 9) to jump to a specific window.

---
Since the feature isn't localized to a single file, I've marked both
change locations with a kind of feature-marker:
  `mycfg-feature:direct-win-focus`
In hope that I'll add more as I add non-single-file features!